move configuration to `./assets/configuration/`
