+++ src/attack_flow_builder/src/assets/configuration/afb.png
