+++ src/attack_flow_builder/src/assets/configuration/ctid.png
