+++ src/attack_flow_builder/src/assets/configuration/ctid_icon_small.png
